requirements.txt
flask
python-dotenv
langchain
langchain_groq
langchain_openai
google-generativeai